requirements.txt
# Python Requirements for Blockchain Encryption System
# Implementation of "Dynamic AES Encryption and Blockchain Key Management"

flask>=2.2.0
gunicorn>=21.2.0

# Core cryptography
pycryptodome>=3.19.0    # AES encryption (GCM mode)
ecdsa>=0.18.0           # ECC (secp256k1) encryption

# Blockchain (Ethereum/Ganache integration)
web3>=6.11.0            # Web3 for Ganache connection

# Standard libraries (included with Python, listed for reference)
# hashlib    - SHA-256 hashing
# json       - JSON serialization
# datetime   - Timestamps
# pathlib    - File path handling
# uuid       - Unique ID generation
